backend/requirements/development.txt
-r base.txt
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
django-debug-toolbar==4.4.6
ipython==8.30.0
